Introduce ZApp, an App with custom R
