+++ core/js/src/main/scala/zio/ZApp.scala
@@ -16,6 +16,17 @@
 
 package zio
 
-trait BootstrapRuntime extends ZBootstrapRuntime[ZEnv] {
-  def environment: ZEnv = ZEnv.Services.live
+trait ZApp[R] extends Runtime[R] {
+
+  /**
+   * The main function of the application, which will be passed the command-line
+   * arguments to the program.
+   */
+  def run(args: List[String]): URIO[R, ExitCode]
+
+  /**
+   * The Scala main function, intended to be called only by the Scala runtime.
+   */
+  final def main(args0: Array[String]): Unit =
+    unsafeRunAsync(run(args0.toList))
 }
